+++ corehq/apps/users/xml.py
@@ -7,8 +7,7 @@ Spec: https://bitbucket.org/javarosa/javarosa/wiki/UserRegistrationAPI
 # this should eventually be harmonized with the other responses, but
 # has been implemented quick and dirty
 from casexml.apps.phone import xml as phone_xml
-from couchforms import xml as couchforms_xml
-from couchforms.xml import ResponseNature
+from couchforms.openrosa_response import ResponseNature, get_response_element
 
 
 def get_response(user, created):
@@ -19,6 +18,6 @@ def get_response(user, created):
         
     nature = ResponseNature.SUBMIT_USER_REGISTERED if created else \
              ResponseNature.SUBMIT_USER_UPDATED
-    response = couchforms_xml.get_response_element(text, nature=nature)
+    response = get_response_element(text, nature=nature)
     response.append(phone_xml.get_registration_element(user.to_casexml_user()))
     return phone_xml.tostring(response)
